Refactor - don't use a client module
