--- aeron-rs/src/client/mod.rs
@@ -1,6 +1,0 @@
-//! Aeron client
-//!
-//! These are the modules necessary to construct a functioning Aeron client
-pub mod cnc_descriptor;
-pub mod concurrent;
-pub mod context;
