+++ aeron-rs/tests/broadcast_receiver.rs
@@ -1,7 +1,7 @@
-use aeron_rs::client::concurrent::broadcast::{
+use aeron_rs::concurrent::broadcast::{
     buffer_descriptor, record_descriptor, BroadcastReceiver,
 };
-use aeron_rs::client::concurrent::AtomicBuffer;
+use aeron_rs::concurrent::AtomicBuffer;
 use aeron_rs::util::bit::align;
 use aeron_rs::util::IndexT;
