+++ aeron-rs/tests/cnc_terminate.rs
@@ -1,8 +1,8 @@
 use aeron_driver_sys::*;
-use aeron_rs::client::cnc_descriptor;
-use aeron_rs::client::cnc_descriptor::MetaDataDefinition;
-use aeron_rs::client::concurrent::ringbuffer::ManyToOneRingBuffer;
-use aeron_rs::client::concurrent::AtomicBuffer;
+use aeron_rs::cnc_descriptor;
+use aeron_rs::cnc_descriptor::MetaDataDefinition;
+use aeron_rs::concurrent::ringbuffer::ManyToOneRingBuffer;
+use aeron_rs::concurrent::AtomicBuffer;
 use aeron_rs::util::IndexT;
 use memmap::MmapOptions;
 use std::ffi::{c_void, CString};
