+++ aeron-rs/tests/many_to_one_ring_buffer.rs
@@ -1,8 +1,8 @@
 /// Tests based on the C++ tests included with Aeron
-use aeron_rs::client::concurrent::ringbuffer::{
+use aeron_rs::concurrent::ringbuffer::{
     buffer_descriptor, record_descriptor, ManyToOneRingBuffer,
 };
-use aeron_rs::client::concurrent::AtomicBuffer;
+use aeron_rs::concurrent::AtomicBuffer;
 use aeron_rs::util::bit::align;
 use aeron_rs::util::IndexT;
 use std::ops::Deref;
